+++ src/assets/index.ts
@@ -1,2 +1,3 @@
 export { default as CartIcon } from "./cart-icon.svg";
 export { default as CartGrayIcon } from "./cart-gray-icon.svg";
+export { default as TrashCanIcon } from "./trashcan-icon.svg";
