+++ src/components/index.ts
@@ -1,5 +1,5 @@
-export { default as Counter } from "./Counter";
-export { default as Header } from "./Header";
-export { default as Product } from "./Product";
-export { default as ProductList } from "./ProductList";
-export { default as Page } from "./Page";
+export { Counter } from "./Counter";
+export { Header } from "./Header";
+export { Page } from "./Page";
+export { ProductList } from "./ProductList";
+export { CartProductList } from "./CartProductList";
